Load only required files
When tzdatabase is loaded at application start,
all file were parsed since it was assumed that the directory has been
manually cleanedup by removing all unnecessary stuff.

Parsing errors in those files caused the application to fail at boot.

Instead of poking with all the unknown files that are regularely added
to the tzdb, only the relevant files are now selected to be parsed.